Cloning trees does not invalidate NodeIDs
With this change, you can insert a node into a tree, clone the tree, and
use the node ID for both trees. Before this change, NodeIDs could
recognize the tree they'd belong to by 'tree_id', but that whole concept
has been removed. This allows for deriving Clone and having it work as
I expected.